LayoutTests/fast/css/counters/counter-increment-tests-expected.txt

PASS layoutTestController.counterValueForElementById('test3') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test4') is '-2147483647'
PASS layoutTestController.counterValueForElementById('test5') is '0'
PASS layoutTestController.counterValueForElementById('test6') is '0'
PASS layoutTestController.counterValueForElementById('test7') is '0'
PASS layoutTestController.counterValueForElementById('test8') is '10'
PASS layoutTestController.counterValueForElementById('test9') is '10'
PASS layoutTestController.counterValueForElementById('test10') is '2147483646'
PASS layoutTestController.counterValueForElementById('test11') is '2147483647'
PASS layoutTestController.counterValueForElementById('test12') is '2147483647'
PASS layoutTestController.counterValueForElementById('test13') is '2147483647'
PASS layoutTestController.counterValueForElementById('test14') is '2'
PASS layoutTestController.counterValueForElementById('test15') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test16') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test17') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test18') is '0'
PASS layoutTestController.counterValueForElementById('test19') is '0'
PASS layoutTestController.counterValueForElementById('test20') is '0'
PASS layoutTestController.counterValueForElementById('test21') is '20'
PASS layoutTestController.counterValueForElementById('test22') is '20'
PASS layoutTestController.counterValueForElementById('test23') is '2147483647'
PASS layoutTestController.counterValueForElementById('test24') is '2147483647'
PASS layoutTestController.counterValueForElementById('test25') is '2147483647'
PASS layoutTestController.counterValueForElementById('test26') is '2147483647'
PASS layoutTestController.counterValueForElementById('test27') is '3'
PASS layoutTestController.counterValueForElementById('test28') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test29') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test30') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test31') is '0'
PASS layoutTestController.counterValueForElementById('test32') is '0'
PASS layoutTestController.counterValueForElementById('test33') is '0'
PASS layoutTestController.counterValueForElementById('test34') is '30'
PASS layoutTestController.counterValueForElementById('test35') is '30'
PASS layoutTestController.counterValueForElementById('test36') is '2147483647'
PASS layoutTestController.counterValueForElementById('test37') is '2147483647'
PASS layoutTestController.counterValueForElementById('test38') is '2147483647'
PASS layoutTestController.counterValueForElementById('test39') is '2147483647'
PASS layoutTestController.counterValueForElementById('test40') is '32'
PASS layoutTestController.counterValueForElementById('test41') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test42') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test43') is '-2147483648'
PASS layoutTestController.counterValueForElementById('test44') is '0'
PASS layoutTestController.counterValueForElementById('test45') is '0'
PASS layoutTestController.counterValueForElementById('test46') is '0'
PASS layoutTestController.counterValueForElementById('test47') is '320'
PASS layoutTestController.counterValueForElementById('test48') is '320'
PASS layoutTestController.counterValueForElementById('test49') is '2147483647'
PASS layoutTestController.counterValueForElementById('test50') is '2147483647'
PASS layoutTestController.counterValueForElementById('test51') is '2147483647'
PASS layoutTestController.counterValueForElementById('test52') is '2147483647'
PASS layoutTestController.counterValueForElementById('test53') is '5'
PASS layoutTestController.counterValueForElementById('none53') is '5'
PASS layoutTestController.counterValueForElementById('test53a') is '10'
PASS layoutTestController.counterValueForElementById('wrapper54') is '5'
PASS layoutTestController.counterValueForElementById('test54') is '10'
PASS layoutTestController.counterValueForElementById('test55') is '1'
PASS layoutTestController.counterValueForElementById('test56') is '-2147483648'
